Expands on generate recurring expenses, makes it work for income as well.
